requirements.txt
aiohttp_jinja2 == 1.5.0
black
connexion[aiohttp,swagger-ui] >= 2.6.0, <3
isort
litgpt
pyupgrade
swagger-ui-bundle == 0.0.9
